requirements.txt
streamlit~=1.44.1
python-dotenv~=1.0.1
requests~=2.32.3
langchain-core~=0.3.51
loguru~=0.7.3
pypinyin~=0.54.0
langchain-community~=0.3.21
langchain~=0.3.23
langchain-openai~=0.3.12
langgraph~=0.3.28
httpx~=0.27.2
uvicorn~=0.32.1
fastapi~=0.115.6